Rework timeline selection components
