+++ res/css/style.css
@@ -329,12 +329,12 @@ body, #root, .profileViewer {
   flex: 1;
 }
 
-.profileViewerHeaderTimeRuler {
+.timelineTimeRuler {
   height: 20px;
   overflow: hidden;
 }
 
-.profileViewerHeaderTimeRuler::after {
+.timelineTimeRuler::after {
   content: '';
   position: absolute;
   top: 20px;
@@ -391,145 +391,6 @@ body, #root, .profileViewer {
   display: block;
   height: 25px;
   width: 100%;
-}
-
-.timeSelectionScrubberHoverIndicator {
-  position: absolute;
-  pointer-events: none;
-  visibility: hidden;
-  top: 0;
-  bottom: 0;
-  width: 1px;
-  background: rgba(0,0,0,0.4);
-  z-index: 1;
-}
-
-.overlay {
-  position: absolute;
-  z-index: 2;
-  display: flex;
-  flex-flow: row nowrap;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  pointer-events: none;
-  margin-left: -5px;
-  padding-left: 5px;
-  overflow: hidden;
-}
-
-.dimmerBefore,
-.dimmerAfter {
-  background: rgba(12, 12, 13, .1);
-  flex-shrink: 0;
-}
-
-.dimmerAfter {
-  flex: 1;
-}
-
-.selectionScrubberGrippy {
-  height: 20px;
-  pointer-events: auto;
-  display: flex;
-  flex-flow: row nowrap;
-}
-
-.grippyRangeStart,
-.grippyRangeEnd {
-  width: 0px;
-  padding: 3px;
-  background: #AAA;
-  border: 1px solid white;
-  margin: 0 -4px;
-  cursor: ew-resize;
-  border-radius: 5px;
-  position: relative;
-  z-index: 3;
-}
-
-.grippyRangeStart:hover,
-.grippyRangeStart.dragging,
-.grippyRangeEnd:hover,
-.grippyRangeEnd.dragging {
-  background: #888;
-}
-
-.grippyMoveRange {
-  flex: 1;
-  cursor: -webkit-grab;
-  cursor: grab;
-}
-
-.grippyMoveRange.dragging {
-  cursor: -webkit-grabbing;
-  cursor: grabbing;
-}
-
-.selectionScrubberWrapper {
-  display: flex;
-  flex-flow: column nowrap;
-}
-
-.selectionScrubberInner {
-  flex: 1;
-  justify-content: center;
-  align-items: center;
-  display: flex;
-  min-width: 0;
-  min-height: 0;
-}
-
-.selectionScrubberRange {
-  top: 20px;
-  position: absolute;
-  padding: 4px 8px;
-  color: #fff;
-  background-color: var(--blue-50);
-  border-radius: 0 0 4px 4px;
-  box-shadow: 0 2px 2px rgba(0, 0, 0, 0.2);
-  pointer-events: none;
-  opacity: 1;
-  transition: opacity 200ms;
-}
-
-.selectionScrubberRange.hidden {
-  opacity: 0;
-}
-
-.selectionScrubberZoomButton {
-  width: 30px;
-  height: 30px;
-  pointer-events: auto;
-  box-sizing: border-box;
-  border-radius: 100%;
-  margin: -15px;
-  position: relative;
-  border: 1px solid rgba(0, 0, 0, 0.2);
-  background: url(../img/svg/zoom-icon.svg) center center no-repeat rgba(255, 255, 255, 0.6);
-  transition: opacity 200ms ease-in-out;
-  will-change: opacity;
-  opacity: 0.5;
-}
-
-.selectionScrubberZoomButton.hidden {
-  opacity: 0.0 !important;
-  pointer-events: none;
-  transition: unset;
-}
-
-.timeline:hover .selectionScrubberZoomButton,
-.selectionScrubberZoomButton:active {
-  opacity: 1.0;
-}
-
-.selectionScrubberZoomButton:hover {
-  background-color: rgba(255, 255, 255, 0.9);
-}
-
-.selectionScrubberZoomButton:active:hover {
-  background-color: rgba(160, 160, 160, 0.6);
 }
 
 .tabBarContainer {
