+++ src/components/timeline/index.css
@@ -2,16 +2,6 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-.timeline {
-  position: relative;
-  margin-left: 149px;
-  border-left: 1px solid var(--grey-30);
-  -moz-user-focus: ignore;
-}
-
-.timeline:hover > .timeSelectionScrubberHoverIndicator {
-  visibility: visible;
-}
 
 .timelineOverflowEdgeIndicatorScrollbox {
   margin: 0 0 0 -150px;
